+++ packages/atomic/src/components.d.ts
@@ -189,6 +189,11 @@ export namespace Components {
          */
         "citation": GeneratedAnswerCitation;
         /**
+          * Option to disable citation anchoring.
+          * @default false
+         */
+        "disableCitationAnchoring": boolean;
+        /**
           * The citation index.
          */
         "index": number;
@@ -677,9 +682,14 @@ export namespace Components {
          */
         "collapsible"?: boolean;
         /**
+          * Option to disable citation anchoring.
+          * @default false
+         */
+        "disableCitationAnchoring"?: boolean;
+        /**
           * A list of fields to include with the citations used to generate the answer.
          */
-        "fieldsToIncludeInCitations": string;
+        "fieldsToIncludeInCitations"?: string;
         /**
           * The maximum height (in rem units) of the answer when collapsed.
          */
@@ -821,9 +831,14 @@ export namespace Components {
          */
         "collapsible"?: boolean;
         /**
+          * Option to disable citation anchoring.
+          * @default false
+         */
+        "disableCitationAnchoring"?: boolean;
+        /**
           * A list of fields to include with the citations used to generate the answer.
          */
-        "fieldsToIncludeInCitations": string;
+        "fieldsToIncludeInCitations"?: string;
         /**
           * The maximum height (in rem units) of the answer when collapsed.
          */
@@ -1579,15 +1594,19 @@ export namespace Components {
     interface AtomicPopover {
     }
     /**
-     * @alpha The `atomic-product-excerpt` component renders the excerpt of a product generated at query time.
+     * @alpha The `atomic-product-description` component renders the description of a product.
      */
-    interface AtomicProductExcerpt {
+    interface AtomicProductDescription {
         /**
-          * Whether the excerpt should be collapsible after being expanded.
+          * The name of the description field to use.
+         */
+        "field": 'ec_description' | 'ec_shortdesc';
+        /**
+          * Whether the description should be collapsible after being expanded.
          */
         "isCollapsible": boolean;
         /**
-          * The number of lines after which the product excerpt should be truncated. A value of "none" will disable truncation.
+          * The number of lines after which the product description should be truncated. A value of "none" will disable truncation.
          */
         "truncateAfter": TruncateAfter;
     }
@@ -1647,37 +1666,6 @@ export namespace Components {
           * Moves to the previous image, when the carousel is activated.
          */
         "previousImage": () => Promise<void>;
-    }
-    /**
-     * @alpha The `atomic-product-link` component automatically transforms a search product title into a clickable link that points to the original item.
-     */
-    interface AtomicProductLink {
-        /**
-          * The [template literal](https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Template_literals) from which to generate the `href` attribute value  The template literal can reference any number of product properties from the parent product. It can also reference the window object.  For example, the following markup generates an `href` value such as `http://uri.com?id=itemTitle`, using the product's `clickUri` and `itemtitle` fields. ```html <atomic-product-link href-template='${clickUri}?id=${permanentId}'></atomic-product-link> ```
-         */
-        "hrefTemplate"?: string;
-    }
-    /**
-     * @alpha The `atomic-product-multi-value-text` component renders the values of a multi-value string field.
-     */
-    interface AtomicProductMultiValueText {
-        /**
-          * The delimiter used to separate values when the field isn't indexed as a multi value field.
-         */
-        "delimiter": string | null;
-        /**
-          * The field that the component should use. The component will try to find this field in the `Product.additionalFields` object unless it finds it in the `Product` object first. Make sure this field is present in the `fieldsToInclude` property of the `atomic-commerce-interface` component.
-         */
-        "field": string;
-        /**
-          * The maximum number of field values to display. If there are _n_ more values than the specified maximum, the last displayed value will be "_n_ more...".
-         */
-        "maxValuesToDisplay": number;
-    }
-    /**
-     * @alpha The `atomic-product-price` component renders the price of a product.
-     */
-    interface AtomicProductPrice {
     }
     /**
      * The `atomic-product-rating` element renders a star rating.
@@ -4128,13 +4116,13 @@ declare global {
         new (): HTMLAtomicPopoverElement;
     };
     /**
-     * @alpha The `atomic-product-excerpt` component renders the excerpt of a product generated at query time.
+     * @alpha The `atomic-product-description` component renders the description of a product.
      */
-    interface HTMLAtomicProductExcerptElement extends Components.AtomicProductExcerpt, HTMLStencilElement {
+    interface HTMLAtomicProductDescriptionElement extends Components.AtomicProductDescription, HTMLStencilElement {
     }
-    var HTMLAtomicProductExcerptElement: {
-        prototype: HTMLAtomicProductExcerptElement;
-        new (): HTMLAtomicProductExcerptElement;
+    var HTMLAtomicProductDescriptionElement: {
+        prototype: HTMLAtomicProductDescriptionElement;
+        new (): HTMLAtomicProductDescriptionElement;
     };
     /**
      * The `atomic-product-field-condition` component takes a list of conditions that, if fulfilled, apply the template in which it's defined.
@@ -4156,33 +4144,6 @@ declare global {
     var HTMLAtomicProductImageElement: {
         prototype: HTMLAtomicProductImageElement;
         new (): HTMLAtomicProductImageElement;
-    };
-    /**
-     * @alpha The `atomic-product-link` component automatically transforms a search product title into a clickable link that points to the original item.
-     */
-    interface HTMLAtomicProductLinkElement extends Components.AtomicProductLink, HTMLStencilElement {
-    }
-    var HTMLAtomicProductLinkElement: {
-        prototype: HTMLAtomicProductLinkElement;
-        new (): HTMLAtomicProductLinkElement;
-    };
-    /**
-     * @alpha The `atomic-product-multi-value-text` component renders the values of a multi-value string field.
-     */
-    interface HTMLAtomicProductMultiValueTextElement extends Components.AtomicProductMultiValueText, HTMLStencilElement {
-    }
-    var HTMLAtomicProductMultiValueTextElement: {
-        prototype: HTMLAtomicProductMultiValueTextElement;
-        new (): HTMLAtomicProductMultiValueTextElement;
-    };
-    /**
-     * @alpha The `atomic-product-price` component renders the price of a product.
-     */
-    interface HTMLAtomicProductPriceElement extends Components.AtomicProductPrice, HTMLStencilElement {
-    }
-    var HTMLAtomicProductPriceElement: {
-        prototype: HTMLAtomicProductPriceElement;
-        new (): HTMLAtomicProductPriceElement;
     };
     /**
      * The `atomic-product-rating` element renders a star rating.
@@ -5264,12 +5225,9 @@ declare global {
         "atomic-numeric-range": HTMLAtomicNumericRangeElement;
         "atomic-pager": HTMLAtomicPagerElement;
         "atomic-popover": HTMLAtomicPopoverElement;
-        "atomic-product-excerpt": HTMLAtomicProductExcerptElement;
+        "atomic-product-description": HTMLAtomicProductDescriptionElement;
         "atomic-product-field-condition": HTMLAtomicProductFieldConditionElement;
         "atomic-product-image": HTMLAtomicProductImageElement;
-        "atomic-product-link": HTMLAtomicProductLinkElement;
-        "atomic-product-multi-value-text": HTMLAtomicProductMultiValueTextElement;
-        "atomic-product-price": HTMLAtomicProductPriceElement;
         "atomic-product-rating": HTMLAtomicProductRatingElement;
         "atomic-product-section-actions": HTMLAtomicProductSectionActionsElement;
         "atomic-product-section-badges": HTMLAtomicProductSectionBadgesElement;
@@ -5480,6 +5438,11 @@ declare namespace LocalJSX {
           * The citation item information.
          */
         "citation": GeneratedAnswerCitation;
+        /**
+          * Option to disable citation anchoring.
+          * @default false
+         */
+        "disableCitationAnchoring"?: boolean;
         /**
           * The citation index.
          */
@@ -5968,6 +5931,11 @@ declare namespace LocalJSX {
          */
         "collapsible"?: boolean;
         /**
+          * Option to disable citation anchoring.
+          * @default false
+         */
+        "disableCitationAnchoring"?: boolean;
+        /**
           * A list of fields to include with the citations used to generate the answer.
          */
         "fieldsToIncludeInCitations"?: string;
@@ -6108,6 +6076,11 @@ declare namespace LocalJSX {
           * @default false
          */
         "collapsible"?: boolean;
+        /**
+          * Option to disable citation anchoring.
+          * @default false
+         */
+        "disableCitationAnchoring"?: boolean;
         /**
           * A list of fields to include with the citations used to generate the answer.
          */
@@ -6831,15 +6804,19 @@ declare namespace LocalJSX {
     interface AtomicPopover {
     }
     /**
-     * @alpha The `atomic-product-excerpt` component renders the excerpt of a product generated at query time.
+     * @alpha The `atomic-product-description` component renders the description of a product.
      */
-    interface AtomicProductExcerpt {
+    interface AtomicProductDescription {
         /**
-          * Whether the excerpt should be collapsible after being expanded.
+          * The name of the description field to use.
+         */
+        "field"?: 'ec_description' | 'ec_shortdesc';
+        /**
+          * Whether the description should be collapsible after being expanded.
          */
         "isCollapsible"?: boolean;
         /**
-          * The number of lines after which the product excerpt should be truncated. A value of "none" will disable truncation.
+          * The number of lines after which the product description should be truncated. A value of "none" will disable truncation.
          */
         "truncateAfter"?: TruncateAfter;
     }
@@ -6886,37 +6863,6 @@ declare namespace LocalJSX {
           * @type {string}
          */
         "imageAltField"?: string;
-    }
-    /**
-     * @alpha The `atomic-product-link` component automatically transforms a search product title into a clickable link that points to the original item.
-     */
-    interface AtomicProductLink {
-        /**
-          * The [template literal](https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Template_literals) from which to generate the `href` attribute value  The template literal can reference any number of product properties from the parent product. It can also reference the window object.  For example, the following markup generates an `href` value such as `http://uri.com?id=itemTitle`, using the product's `clickUri` and `itemtitle` fields. ```html <atomic-product-link href-template='${clickUri}?id=${permanentId}'></atomic-product-link> ```
-         */
-        "hrefTemplate"?: string;
-    }
-    /**
-     * @alpha The `atomic-product-multi-value-text` component renders the values of a multi-value string field.
-     */
-    interface AtomicProductMultiValueText {
-        /**
-          * The delimiter used to separate values when the field isn't indexed as a multi value field.
-         */
-        "delimiter"?: string | null;
-        /**
-          * The field that the component should use. The component will try to find this field in the `Product.additionalFields` object unless it finds it in the `Product` object first. Make sure this field is present in the `fieldsToInclude` property of the `atomic-commerce-interface` component.
-         */
-        "field": string;
-        /**
-          * The maximum number of field values to display. If there are _n_ more values than the specified maximum, the last displayed value will be "_n_ more...".
-         */
-        "maxValuesToDisplay"?: number;
-    }
-    /**
-     * @alpha The `atomic-product-price` component renders the price of a product.
-     */
-    interface AtomicProductPrice {
     }
     /**
      * The `atomic-product-rating` element renders a star rating.
@@ -8540,12 +8486,9 @@ declare namespace LocalJSX {
         "atomic-numeric-range": AtomicNumericRange;
         "atomic-pager": AtomicPager;
         "atomic-popover": AtomicPopover;
-        "atomic-product-excerpt": AtomicProductExcerpt;
+        "atomic-product-description": AtomicProductDescription;
         "atomic-product-field-condition": AtomicProductFieldCondition;
         "atomic-product-image": AtomicProductImage;
-        "atomic-product-link": AtomicProductLink;
-        "atomic-product-multi-value-text": AtomicProductMultiValueText;
-        "atomic-product-price": AtomicProductPrice;
         "atomic-product-rating": AtomicProductRating;
         "atomic-product-section-actions": AtomicProductSectionActions;
         "atomic-product-section-badges": AtomicProductSectionBadges;
@@ -8862,9 +8805,9 @@ declare module "@stencil/core" {
              */
             "atomic-popover": LocalJSX.AtomicPopover & JSXBase.HTMLAttributes<HTMLAtomicPopoverElement>;
             /**
-             * @alpha The `atomic-product-excerpt` component renders the excerpt of a product generated at query time.
+             * @alpha The `atomic-product-description` component renders the description of a product.
              */
-            "atomic-product-excerpt": LocalJSX.AtomicProductExcerpt & JSXBase.HTMLAttributes<HTMLAtomicProductExcerptElement>;
+            "atomic-product-description": LocalJSX.AtomicProductDescription & JSXBase.HTMLAttributes<HTMLAtomicProductDescriptionElement>;
             /**
              * The `atomic-product-field-condition` component takes a list of conditions that, if fulfilled, apply the template in which it's defined.
              * The condition properties can be based on any top-level product property of the `product` object, not restricted to fields (e.g., `ec_name`).
@@ -8876,18 +8819,6 @@ declare module "@stencil/core" {
              * @alpha 
              */
             "atomic-product-image": LocalJSX.AtomicProductImage & JSXBase.HTMLAttributes<HTMLAtomicProductImageElement>;
-            /**
-             * @alpha The `atomic-product-link` component automatically transforms a search product title into a clickable link that points to the original item.
-             */
-            "atomic-product-link": LocalJSX.AtomicProductLink & JSXBase.HTMLAttributes<HTMLAtomicProductLinkElement>;
-            /**
-             * @alpha The `atomic-product-multi-value-text` component renders the values of a multi-value string field.
-             */
-            "atomic-product-multi-value-text": LocalJSX.AtomicProductMultiValueText & JSXBase.HTMLAttributes<HTMLAtomicProductMultiValueTextElement>;
-            /**
-             * @alpha The `atomic-product-price` component renders the price of a product.
-             */
-            "atomic-product-price": LocalJSX.AtomicProductPrice & JSXBase.HTMLAttributes<HTMLAtomicProductPriceElement>;
             /**
              * The `atomic-product-rating` element renders a star rating.
              * @alpha