+++ packages/atomic/src/components/commerce/lazy-index.ts
@@ -76,12 +76,20 @@ export default {
     await import('./atomic-product/atomic-product.js'),
   'atomic-product-children': async () =>
     await import('./atomic-product-children/atomic-product-children.js'),
-  'atomic-product-description': async () =>
-    await import('./atomic-product-description/atomic-product-description.js'),
+  'atomic-product-excerpt': async () =>
+    await import('./atomic-product-excerpt/atomic-product-excerpt.js'),
+  'atomic-product-link': async () =>
+    await import('./atomic-product-link/atomic-product-link.js'),
+  'atomic-product-multi-value-text': async () =>
+    await import(
+      './atomic-product-multi-value-text/atomic-product-multi-value-text.js'
+    ),
   'atomic-product-numeric-field-value': async () =>
     await import(
       './atomic-product-numeric-field-value/atomic-product-numeric-field-value.js'
     ),
+  'atomic-product-price': async () =>
+    await import('./atomic-product-price/atomic-product-price.js'),
   'atomic-product-template': async () =>
     await import('./atomic-product-template/atomic-product-template.js'),
   'atomic-product-text': async () =>