+++ packages/atomic/src/components/insight/atomic-insight-generated-answer/atomic-insight-generated-answer.tsx
@@ -73,7 +73,7 @@ export class AtomicInsightGeneratedAnswer
   private readonly DEFAULT_COLLAPSED_HEIGHT = 16;
   private readonly MAX_COLLAPSED_HEIGHT = 32;
   private readonly MIN_COLLAPSED_HEIGHT = 9;
-  private readonly DEFAULT_FIELDS_TO_INCLUDE_IN_CITATIONS = 'filetype';
+  private readonly REQUIRED_FIELDS_TO_INCLUDE_IN_CITATIONS = ['filetype'];
 
   @BindStateToController('generatedAnswer', {
     onUpdateCallbackMethod: 'onGeneratedAnswerStateUpdate',
@@ -123,8 +123,13 @@ export class AtomicInsightGeneratedAnswer
   /**
    * A list of fields to include with the citations used to generate the answer.
    */
-  @Prop() fieldsToIncludeInCitations =
-    this.DEFAULT_FIELDS_TO_INCLUDE_IN_CITATIONS;
+  @Prop() fieldsToIncludeInCitations?: string;
+
+  /**
+   * Option to disable citation anchoring.
+   * @default false
+   */
+  @Prop() disableCitationAnchoring?: boolean;
 
   @AriaLiveRegion('generated-answer')
   protected ariaMessage!: string;
@@ -137,6 +142,7 @@ export class AtomicInsightGeneratedAnswer
       host: this.host,
       withToggle: this.withToggle,
       collapsible: this.collapsible,
+      disableCitationAnchoring: this.disableCitationAnchoring,
       getGeneratedAnswer: () => this.generatedAnswer,
       getGeneratedAnswerState: () => this.generatedAnswerState,
       getSearchStatusState: () => this.searchStatusState,
@@ -258,10 +264,11 @@ export class AtomicInsightGeneratedAnswer
   }
 
   private getCitationFields() {
-    return this.fieldsToIncludeInCitations
-      ?.split(',')
+    return (this.fieldsToIncludeInCitations ?? '')
+      .split(',')
       .map((field) => field.trim())
-      .filter((field) => field.length > 0);
+      .filter((field) => field.length > 0)
+      .concat(this.REQUIRED_FIELDS_TO_INCLUDE_IN_CITATIONS)
   }
 
   private validateMaxCollapsedHeight(): number {